Refactor package name with pojo to differentiate dto and pojo subpackages from model package
