+++ src/main/java/com/medicoLaboSolutions/backPatient/service/PatientService.java
@@ -1,13 +1,11 @@
 package com.medicoLaboSolutions.backPatient.service;
 
 import com.medicoLaboSolutions.backPatient.exceptions.PatientNotFoundException;
-import com.medicoLaboSolutions.backPatient.model.Patient;
+import com.medicoLaboSolutions.backPatient.model.pojo.Patient;
 import com.medicoLaboSolutions.backPatient.model.dto.PatientDTO;
 import com.medicoLaboSolutions.backPatient.repository.PatientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 public class PatientService {
